--- src/main/java/com/example/psp/domain/repositories/transaction/TransactionRepository.java
@@ -1,9 +1,0 @@
-package com.example.psp.domain.repositories.transaction;
-
-import com.example.psp.domain.entities.Transaction;
-
-public interface TransactionRepository {
-
-    void save(Transaction transaction);
-    void update(Transaction transaction);
-}
