--- src/main/java/com/example/psp/domain/services/acquirer/AcquirerRouter.java
@@ -1,8 +1,0 @@
-package com.example.psp.domain.services.acquirer;
-
-import com.example.psp.domain.valueobjects.CardDetails;
-
-public interface AcquirerRouter {
-
-    Acquirer getAcquirer(CardDetails cardDetails);
-}
