--- src/main/java/com/example/psp/domain/services/transaction/TransactionService.java
@@ -1,8 +1,0 @@
-package com.example.psp.domain.services.transaction;
-
-import com.example.psp.domain.entities.Transaction;
-
-public interface TransactionService {
-
-    PaymentResponse processPayment(PaymentRequest paymentDetails);
-}
