+++ src/main/java/com/example/psp/service/acquirer/ports/Acquirer.java
@@ -1,13 +1,14 @@
-package com.example.psp.domain.services.acquirer;
+package com.example.psp.service.acquirer.ports;
 
 import com.example.psp.domain.enums.AcquirerDecision;
 import com.example.psp.domain.enums.AcquirerType;
 import com.example.psp.domain.valueobjects.CardDetails;
 import com.example.psp.domain.valueobjects.Money;
+import reactor.core.publisher.Mono;
 
 public interface Acquirer {
 
-    AcquirerDecision authorizeTransaction(CardDetails cardDetails, Money money);
+    Mono<AcquirerDecision> authorizeTransaction(CardDetails cardDetails, Money money);
 
     AcquirerType getType();
