+++ src/main/java/com/example/psp/service/transaction/ports/TransactionService.java
@@ -1,0 +1,8 @@
+package com.example.psp.service.transaction.ports;
+
+import reactor.core.publisher.Mono;
+
+public interface TransactionService {
+
+    Mono<PaymentResponse> processPayment(PaymentRequest paymentDetails);
+}
